refactor sources, packages, exports
