+++ jsrc/csvClient.sc
@@ -1,10 +1,9 @@
 #!/usr/bin/env -S scala-cli shebang -Wunused:imports -Wunused:locals -deprecation
 
-//> using dep org.vastblue:uni_3:0.6.1
+//> using dep org.vastblue:uni_3:0.6.2
 
 import java.nio.charset.StandardCharsets
 import uni.*
-import uni.fs.*
 import uni.io.*
 
 object CsvClient {
