+++ jsrc/csvClientRows.sc
@@ -1,9 +1,9 @@
 #!/usr/bin/env -S scala-cli shebang -Wunused:imports -Wunused:locals -deprecation
 
-//> using dep org.vastblue:uni_3:0.6.1
+//> using dep org.vastblue:uni_3:0.6.2
 
 import uni.*
-import uni.fs.*
+//import uni.fs.*
 import uni.io.*
 
 object CsvClientRows {
