+++ src/main/scala/apps/DelimiterCheck.scala
@@ -4,7 +4,6 @@ package uni
 //> using dep org.vastblue:uni_3:0.6.1
 
 import uni.*
-import uni.fs.*
 import uni.io.*
 
 object DelimiterCheck {
