+++ src/main/scala/uni/PathsUtils.scala
@@ -3,18 +3,14 @@ package uni
 
 import java.io.{File as JFile}
 import java.io.{BufferedReader, InputStreamReader}
-import java.nio.charset.{Charset, StandardCharsets}
-import java.nio.file.{Path, Files, Paths as JPaths, StandardCopyOption}
+import java.nio.charset.{StandardCharsets}
+import java.nio.file.{Path, Files, Paths as JPaths}
 import java.util.Locale
 import scala.collection.immutable.SortedMap
 import scala.sys.process.*
 import scala.util.Try
 import scala.util.Properties
-import scala.jdk.CollectionConverters.*
-import scala.math.BigDecimal.RoundingMode
-import StandardCharsets.{UTF_8, ISO_8859_1 as Latin1}
-import Internals.*
-import uni.fs.*
+import uni.ext.*
 
 export scala.util.Properties.{isWin, isMac, isLinux}
 export Proc.{spawn, call, lazyLines}
@@ -153,7 +149,7 @@ object Proc {
 lazy val pwd: Path = JPaths.get(config.userdir)
 
 object Internals {
-  import fs.*
+  import ext.*
 
   def realpathWindows(path: String): String = {
     if (!isWin) {
@@ -570,290 +566,93 @@ private inline def isDriveLetterPath(s: String): Boolean = {
   }
 }
 
-/** Extension methods */
-object stringExts {
-  extension (str: String) {
-    def path: Path = Paths.get(str)
-    def posx: String = normalizePosix(str)
-    def posix: String = posixAbs(str)
-
-    def startsWithIgnoreCase(prefix: String): Boolean = startsWithUncased(str, prefix)
-    def stripPrefixIgnoreCase(prefix: String): String = stripPrefixUncased(str, prefix)
-    def stripPrefix(prefix: String): String =
-      if str.startsWith(prefix) then str.substring(prefix.length)
-      else str
-
-    def local: String = {
-      val forward = normalizePosix(str)
-      val isPosix = forward.startsWith("/") //|| !forward.matches("(?i)^[a-z]:/.*")
-      if (!isWin || !isPosix) {
-        str
-      } else {
-        config.posix2win.collectFirst {
-          case (posix, winSeq) if posix != "/" && str.startsWithIgnoreCase(posix) =>
-            s"${winSeq.head}${str.stripPrefixIgnoreCase(posix)}"
-        }.getOrElse {
-          s"${config.msysRoot}$str"
-        }
-      }
-    }
-  }
-
-  def startsWithUncased(str: String, prefix: String): Boolean = {
-    str.regionMatches(true, 0, prefix, 0, prefix.length)
-  }
-
-  def stripPrefixUncased(str: String, prefix: String): String = {
-    if startsWithUncased(str, prefix) then str.substring(prefix.length) else str
-  }
-
-}
-
-object pathExts {
-  extension (p: Path) {
-    def exists: Boolean = Files.exists(p)
-    def isDirectory: Boolean = Files.isDirectory(p)
-    def isFile: Boolean = Files.isRegularFile(p)
-    def name: String = p.getFileName.toString
-    def basename: String = {
-      val n = p.getFileName.toString
-      val i = n.lastIndexOf('.')
-      if i == -1 then n else n.substring(0, i)
-    }
-    def relativePath: Path = relativePathToCwd(p)
-    def relpath: String = {
-      val rp: Path = relativePathToCwd(p)
-      standardizePath(rp)
-    }
-
-    def abs: String =
-      if (java.nio.file.Files.exists(p))
-        normalizePosix(p.toAbsolutePath.normalize.toString)
-      else
-        normalizePosix(p.normalize.toString)
-
-    def firstline: String = lines.nextOption.getOrElse("")
-    def lines: Iterator[String] = {
-      try {
-        Files.readAllLines(p, UTF_8).asScala.iterator
-      } catch {
-      case m: java.nio.charset.MalformedInputException =>
-         Files.readAllLines(p, Latin1).asScala.iterator
-      }
-    }
-    def csvRows: Iterator[Seq[String]] = {
-      uni.io.FastCsv.rowsAsync(p)
-    }
-    def csvRows(onRow: Seq[String] => Unit): Unit = {
-      uni.io.FastCsv.eachRow(p){ (row: Seq[String]) =>
-        onRow(row)
-      }
-    }
-    def lines(charset: Charset = UTF_8): Iterator[String] = {
-      scala.io.Source.fromFile(p.toFile).getLines
-    }
-    def contentAsString(charset: Charset = UTF_8): String = Files.readString(p, charset)
-    def contentAsString: String = {
-      try {
-        Files.readString(p, UTF_8)
-      } catch {
-      case m: java.nio.charset.MalformedInputException =>
-        Files.readString(p, Latin1)
-      }
-    }
-    def isSymbolicLink: Boolean = Files.isSymbolicLink(p)
-    def stdpath: String = standardizePath(p)
-    def getParentNonNull: Path = Option(p.getParent).getOrElse(p) // dirname convention: `dirname :/` == /
-    def getParentPath: Path = Option(p.getParent).getOrElse(p.toAbsolutePath.normalize.getParent)
-    def parent: Path = p.toAbsolutePath.getParent
-    def isSameFile(other: Any): Boolean = {
-      try {
-        other match {
-          case otherPath: Path =>
-            sameFileTest(p, otherPath)
-          case _ =>
-            false
-        }
-      } catch {
-        case _: Exception => false
-      }
-    }
-    def localpath: String = normalizePosix(p.toString)
-
-    def dospath: String = {
-      val pstr = p.toString
-      pstr match {
-        case "." => "."
-        case s if !isWin || s.length > 2 =>
-          s
-        case s if s.endsWith(":") =>
-          if rootDrives.contains(s.toUpperCase) then
-            p.toAbsolutePath.toString
-          else
-            p.toFile.getAbsolutePath.toString
-        case s =>
-          s
-      }
-    }
-    def files: Iterator[JFile] = Option(p.toFile.listFiles) match {
-      case Some(arr) => arr.iterator
-      case None      => Iterator.empty
-    }
-    def paths: Iterator[Path] = files.map(_.toPath)
-    def posx: String = {
-      if (p == null) {
-        hook += 1
-      }
-      normalizePosix(p.toString)
-    }
-    def posix: String = posixAbs(p.toString)
-
-    def local: String = {
-      normalizePosix(p.toString)
-    }
-    def dotsuffix: String = {
-      val name = p.getFileName.toString
-      val idx  = name.lastIndexOf('.')
-      if idx > 0 then name.substring(idx) else ""
-    }
-    def suffix: String = {
-      val ext = dotsuffix
-      if ext.nonEmpty then ext.drop(1) else ""
-    }
-    def newerThan(other: Path): Boolean = {
-      p.isFile && other.isFile && other.lastModified > p.lastModified
-    }
-    def lastModified: Long = p.toFile.lastModified
-    def lastModMillisAgo: Long = System.currentTimeMillis - p.toFile.lastModified
-
-    def lastModSecondsAgo: Double   = lastModMillisAgo / 1000.0
-    def lastModMinutesAgo: Double   = fs.round(lastModSecondsAgo / 60.0)
-    def lastModHoursAgo: Double     = fs.round(lastModMinutesAgo / 60.0)
-    def lastModDaysAgo: Double      = fs.round(lastModHoursAgo / 24.0)
-
-    def lastModSeconds: Double = lastModSecondsAgo // alias
-    def lastModMinutes: Double = lastModMinutesAgo // alias
-    def lastModHours: Double   = lastModHoursAgo   // alias
-    def lastModDays: Double    = lastModDaysAgo    // alias
-
-    def lastModifiedYMD: String = {
-      def lastModified = p.toFile.lastModified
-      val date         = new java.util.Date(lastModified)
-      val ymdHms       = new java.text.SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
-      ymdHms.format(date)
-    }
-
-    /** Copy this Path to the given destination Path.
-      * By default overwrites if the target exists.
-      */
-    def copyTo(dest: Path, overwrite: Boolean = true, copyAttributes: Boolean = false): Path = {
-      val options =
-        if (overwrite && copyAttributes)
-          Array(StandardCopyOption.REPLACE_EXISTING, StandardCopyOption.COPY_ATTRIBUTES)
-        else if (overwrite)
-          Array(StandardCopyOption.REPLACE_EXISTING)
-        else if (copyAttributes)
-          Array(StandardCopyOption.COPY_ATTRIBUTES)
-        else
-          Array.empty[StandardCopyOption]
-
-      Files.copy(p, dest, options*)
-      dest
-    }
-    /** Recursively iterate all files and directories under this Path. */
-    def pathsTree: Iterator[Path] =
-      if Files.exists(p) then
-        Files.walk(p).iterator().asScala
-      else
-        Iterator.empty
-  }
-}
-
-object fs {
-  export pathExts.*
-  export stringExts.*
-
-  def round(number: Double, scale: Int = 6): Double =
-    BigDecimal(number).setScale(scale, RoundingMode.HALF_UP).toDouble
 
 
-  def isValidWindowsPath(s: String): Boolean = {
-    val t = s.trim
-
-    val driveAbs   = "^[A-Za-z]:[\\\\/].*".r
-    val unc        = "^\\\\\\\\[^\\\\]+\\\\[^\\\\]+.*".r
-    val device     = "^\\\\\\\\[.?]\\\\.*".r
-    val winRel     = "^[^/]*\\\\.*".r
-    val embedded   = ".*[A-Za-z]:[\\\\/].*".r
-
-    val valid = t match
-      case driveAbs()  => true
-      case unc()       => true
-      case device()    => true
-      case winRel()    => true
-      case embedded()  => true
-      case _           => false
-
-    valid // single exit
-  }
-
-  def isValidMsysPath(s: String): Boolean = {
-    val t = s.trim
-
-    val tildeUser = "^~[A-Za-z0-9._-]+/.*".r
-    val msysDrive = "^/[A-Za-z]/.*".r
-    val optValue  = """.*=[^=]*[/~].*""".r
-
-    // 1. Tilde paths
-    val valid = if t == "~" then true
-    else if t.startsWith("~/") then true
-    else if tildeUser.matches(t) then true
-
-    // 2. Rooted POSIX paths
-    else if t.startsWith("/") && !t.startsWith("//") then true
-
-    // 3. MSYS drive paths (/c/foo)
-    else if msysDrive.matches(t) then true
-
-    // 4. Relative POSIX paths containing '/'
-    else if t.contains("/") && !isValidWindowsPath(t) then true
-
-    // 5. Option-value forms (--foo=/bar, --foo=~/baz)
-    else if optValue.matches(t) && !isValidWindowsPath(t) then true
-
-    else false
-
-    valid // single exit
-  }
-
-  import java.io.{FileWriter, OutputStreamWriter, PrintWriter}
-  def withFileWriter(p: Path, charsetName: String = "UTF-8", append: Boolean = false)(func: PrintWriter => Any): Unit = {
-    val jfile  = p.toFile
-    val lcname = jfile.getName.toLowerCase(Locale.ROOT)
-    if (lcname != "stdout") {
-      Option(jfile.getParentFile) match {
-      case Some(parent) =>
-        if (!parent.exists) {
-          throw new IllegalArgumentException(s"parent directory not found [${parent}]")
-        }
-      case None =>
-        throw new IllegalArgumentException(s"no parent directory")
-      }
-    }
-    val writer = lcname match {
-    case "stdout" =>
-      new PrintWriter(new OutputStreamWriter(System.out, charsetName), true)
-    case _ =>
-      new PrintWriter(new FileWriter(jfile, append))
-    }
-    try {
-      val _: Any = func(writer)
-    } finally {
-      writer.flush()
-      if (lcname != "stdout") {
-        // don't close stdout!
-        writer.close()
-      }
-    }
-  }
-}
+//object fs {
+//  //export pathExts.*
+//  //export stringExts.*
+//
+//  def round(number: Double, scale: Int = 6): Double =
+//    BigDecimal(number).setScale(scale, RoundingMode.HALF_UP).toDouble
+//
+//
+//  def isValidWindowsPath(s: String): Boolean = {
+//    val t = s.trim
+//
+//    val driveAbs   = "^[A-Za-z]:[\\\\/].*".r
+//    val unc        = "^\\\\\\\\[^\\\\]+\\\\[^\\\\]+.*".r
+//    val device     = "^\\\\\\\\[.?]\\\\.*".r
+//    val winRel     = "^[^/]*\\\\.*".r
+//    val embedded   = ".*[A-Za-z]:[\\\\/].*".r
+//
+//    val valid = t match
+//      case driveAbs()  => true
+//      case unc()       => true
+//      case device()    => true
+//      case winRel()    => true
+//      case embedded()  => true
+//      case _           => false
+//
+//    valid // single exit
+//  }
+//
+//  def isValidMsysPath(s: String): Boolean = {
+//    val t = s.trim
+//
+//    val tildeUser = "^~[A-Za-z0-9._-]+/.*".r
+//    val msysDrive = "^/[A-Za-z]/.*".r
+//    val optValue  = """.*=[^=]*[/~].*""".r
+//
+//    // 1. Tilde paths
+//    val valid = if t == "~" then true
+//    else if t.startsWith("~/") then true
+//    else if tildeUser.matches(t) then true
+//
+//    // 2. Rooted POSIX paths
+//    else if t.startsWith("/") && !t.startsWith("//") then true
+//
+//    // 3. MSYS drive paths (/c/foo)
+//    else if msysDrive.matches(t) then true
+//
+//    // 4. Relative POSIX paths containing '/'
+//    else if t.contains("/") && !isValidWindowsPath(t) then true
+//
+//    // 5. Option-value forms (--foo=/bar, --foo=~/baz)
+//    else if optValue.matches(t) && !isValidWindowsPath(t) then true
+//
+//    else false
+//
+//    valid // single exit
+//  }
+//
+//  import java.io.{FileWriter, OutputStreamWriter, PrintWriter}
+//  def withFileWriter(p: Path, charsetName: String = "UTF-8", append: Boolean = false)(func: PrintWriter => Any): Unit = {
+//    val jfile  = p.toFile
+//    val lcname = jfile.getName.toLowerCase(Locale.ROOT)
+//    if (lcname != "stdout") {
+//      Option(jfile.getParentFile) match {
+//      case Some(parent) =>
+//        if (!parent.exists) {
+//          throw new IllegalArgumentException(s"parent directory not found [${parent}]")
+//        }
+//      case None =>
+//        throw new IllegalArgumentException(s"no parent directory")
+//      }
+//    }
+//    val writer = lcname match {
+//    case "stdout" =>
+//      new PrintWriter(new OutputStreamWriter(System.out, charsetName), true)
+//    case _ =>
+//      new PrintWriter(new FileWriter(jfile, append))
+//    }
+//    try {
+//      val _: Any = func(writer)
+//    } finally {
+//      writer.flush()
+//      if (lcname != "stdout") {
+//        // don't close stdout!
+//        writer.close()
+//      }
+//    }
+//  }
+//}
