+++ src/main/scala/uni/cli/package.scala
@@ -1,0 +1,6 @@
+package uni
+
+package object cli:
+  // Public API
+  export ArgsParser.{showUsage, eachArg, thisArg, consumeNext, peekNext, nextInt, nextLong, nextDouble}
+
