+++ src/main/scala/uni/ext/Ext.scala
@@ -1,0 +1,7 @@
+package uni.ext
+
+object Ext:
+  export pathExts.*
+  export stringExts.*
+  export helpers.*
+
