+++ src/main/scala/uni/ext/package.scala
@@ -1,0 +1,4 @@
+package uni
+
+package object ext:
+  export Ext.*
