+++ src/main/scala/uni/package.scala
@@ -1,0 +1,6 @@
+package uni
+
+export cli.ArgsParser.*
+export ext.pathExts.*
+export ext.stringExts.*
+export ext.helpers.*
