+++ src/test/scala/uni/PosixFmtSuite.scala
@@ -3,6 +3,7 @@ package uni
 import munit.FunSuite
 import uni.TestUtils.*
 import java.nio.file.{Paths as JPaths}
+import uni.ext.*
 
 final class PosixFmtSuite extends FunSuite {
 // uncomment the next 2 lines to disable timeout
@@ -16,7 +17,6 @@ final class PosixFmtSuite extends FunSuite {
   }
 
   // import the extension under test
-  import uni.fs.*
   var testsTotal = 0
   var runsTotal = 0
